add routines/options to deep-memtest and to sweep RTT options
These are meant to allow oscilloscope/freeze-spray characterization
of the bus performance on a running tally basis. Use the ddrdebug
option to build and access.

The only impact on mainline should be the addition of a print-out
of temperature at boot to the BIOS, so that units that fail to
calibrate can report their die temp for future diagnostic info.